HalfTensor to FloatTensor

+++ guided_diffusion/unet2.py
@@ -1165,9 +1165,8 @@ class EncoderUNetModel(nn.Module):
         emb = self.time_embed(timestep_embedding(timesteps, self.model_channels))
 
         results = []
-        h = x.type(th.cuda.HalfTensor)
-        
-        # h = x.to(torch.float16)
+        # h = x.type(th.cuda.HalfTensor)
+        h = x.type(th.cuda.FloatTensor)
 
         for module in self.input_blocks:
             h = module(h, emb)
